Remove component status page in favor of on-page statuses.
