--- docs/content/component-status/index.mdx
@@ -1,85 +1,0 @@
----
-name: Component Status
-route: /component-status
----
-
-import { Deprecated, InReview, Stable } from './components/Statuses';
-
-# Component Status
-
-<Stable/> - Component's name and API are considered stable. No foreseen breaking changes.
-
-<br />
-
-<InReview/> - Component is being reviewed as to its name, API, or whether it should be exposed in the UI library.
-
-<br />
-
-<Deprecated/> - Component will be removed in the next major version.
-
-## Components
-
-| Component          | Added             | Status        |
-| ------------------ | ----------------- | ------------- |
-| Badge              | development-0.0.1 | <Stable/>     |
-| Blockquote         | development-0.0.2 | <Stable/>     |
-| Button             | development-0.0.1 | <Stable/>     |
-| Calendar           | development-0.0.1 | <InReview/>   |
-| Card               | development-0.0.1 | <Stable/>     |
-| Checkbox           | development-0.0.1 | <Stable/>     |
-| Heading            | development-0.0.1 | <Stable/>     |
-| Hr                 | development-0.0.1 | <Stable/>     |
-| Image              | development-0.0.1 | <Stable/>     |
-| InlineNotification | development-0.0.1 | <InReview/>   |
-| List               | development-0.0.1 | <Stable/>     |
-| Markdown           | development-0.0.1 | <Stable/>     |
-| Message            | development-0.0.1 | <InReview/>   |
-| Modal              | development-0.0.1 | <Stable/>     |
-| Picture            | development-0.0.1 | <Deprecated/> |
-| ProgressBar        | development-0.0.1 | <Stable/>     |
-| Select             | development-0.0.1 | <Stable/>     |
-| Selector           | development-0.0.1 | <Stable/>     |
-| Subheading         | development-0.0.1 | <Stable/>     |
-| Table              | development-0.0.1 | <Stable/>     |
-| Tag                | development-0.0.1 | <Stable/>     |
-| Text               | development-0.0.1 | <Stable/>     |
-| TextArea           | development-0.0.1 | <Stable/>     |
-| Toggle             | development-0.0.1 | <Stable/>     |
-| Tooltip            | development-0.0.1 | <InReview/>   |
-
-## Grid
-
-| Component | Added             | Status    |
-| --------- | ----------------- | --------- |
-| Col       | development-0.0.1 | <Stable/> |
-| Grid      | development-0.0.1 | <Stable/> |
-| Row       | development-0.0.1 | <Stable/> |
-
-## Patterns
-
-| Component         | Added             | Status      |
-| ----------------- | ----------------- | ----------- |
-| AutoCompleteInput | development-0.0.1 | <InReview/> |
-| AutoCompleteTags  | development-0.0.1 | <InReview/> |
-| CreditCardDetails | development-0.0.1 | <InReview/> |
-| CurrencyInput     | development-0.0.1 | <InReview/> |
-| SearchInput       | development-0.0.1 | <InReview/> |
-| SvgButton         | development-0.0.1 | <InReview/> |
-
-## Utilities
-
-| Component          | Added             | Status      |
-| ------------------ | ----------------- | ----------- |
-| CardSchemes        | development-0.0.1 | <InReview/> |
-| CloseButton        | development-0.0.1 | <InReview/> |
-| Hamburger          | development-0.0.1 | <InReview/> |
-| InfoIcon           | development-0.0.1 | <InReview/> |
-| InlineElements     | development-0.0.1 | <InReview/> |
-| MaskedInput        | development-0.0.1 | <InReview/> |
-| NotificationBanner | development-0.0.1 | <InReview/> |
-| NotificationList   | development-0.0.1 | <InReview/> |
-| PaymentMethodIcon  | development-0.0.1 | <InReview/> |
-| Popover            | development-0.0.1 | <InReview/> |
-| RestrictedInput    | development-0.0.1 | <InReview/> |
-| Spacing            | development-0.0.1 | <InReview/> |
-| Spinner            | development-0.0.1 | <InReview/> |
