+++ docs/content/components/badge.mdx
@@ -7,8 +7,11 @@ menu: Components
 import { Playground } from 'docz';
 import { Fragment } from 'react';
 import Badge from '../../../src/components/Badge/Badge';
+import Status from '../../utils/Statuses';
 
 # Badge
+
+<Status.Stable/>
 
 Badges are non-actionable elements in the UI, used to indicate status of an object.
